+++ supabase/functions/evaluate/evaluators/slider.ts
